### 2017-01-13 #163 @whedon
[PRE REVIEW]: nse: Computation of Numerical Standard Errors in R
**Submitting author:** @ArdiaD (<a href="http://orcid.org/0000-0003-2823-782X">David Ardia</a>)
**Repository:** <a href="https://github.com/keblu/nse" target ="_blank">https://github.com/keblu/nse</a>
**Version:** v1-00.16
