### 2017-01-20 dbabb72
+ doc updated + CRAN README added
