+++ README CRAN.md
@@ -1,0 +1,11 @@
+`nse` is an R package for computing the numerical standard error (NSE), an estimate 
+of the standard deviation of a simulation result, if the simulation experiment were to be repeated 
+many times. The package provides a set of wrappers around several R packages, which give access to 
+more than thirty NSE estimators, including batch means estimators, initial sequence 
+estimators, spectrum at zero estimators, heteroskedasticity and autocorrelation 
+consistent (HAC) kernel estimators and bootstrap estimators. 
+
+The full set of methods available in the package is summarized in [Ardia et al. (2016)](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2741587) together with 
+several examples of applications of NSE in econometrics and finance.
+
+The latest version of the package is available at 'https://github.com/keblu/nse'.
