### 2017-01-24 89b195d
Fixed typo in description
