### 2017-01-24 f02be4d
README updated with nice tags
