### 2017-01-24 4a7c0d6
Update readme to add CRAN submission
