### 2017-01-31 #163 @arfon
> @arfon We have released version v1-00.16, and the version we asked for the review was v1-00.15. How can we ensure that the review will be on the current version number, v1-00.16?

Thanks for the heads up - I've updated the version at the top of this thread to track this.
