### 2017-01-31 #163 @ArdiaD
Btw @arfon, maybe @braverock would also be willing to review this package.
