### 2017-01-31 #163 @braverock
@arfon I agree, I am familiar with this work, and have published similar work in other places. I am willing to review the NSE package.
