### 2017-02-01 #172 @whedon
[REVIEW]: nse: Computation of Numerical Standard Errors in R
**Submitting author:** @ArdiaD (<a href="http://orcid.org/0000-0003-2823-782X">David Ardia</a>)
**Repository:** <a href="https://github.com/keblu/nse" target ="_blank">https://github.com/keblu/nse</a>
**Version:** v1-00.15

- [x] **Version:** Does the release version given match the GitHub release (v1-00.15)?
- [x] **Authorship:** Has the submitting author (ArdiaD) made major contributions to the software?
