### 2017-02-01 #163 @arfon
> @arfon I agree, I am familiar with this work, and have published similar work in other places. I am willing to review the NSE package.

@braverock - many thanks. Please head over to #172 where the review should take place.
