### 2017-02-01 #172 @arfon
@braverock - please take a look at the reviewer guidelines here (http://joss.theoj.org/about#reviewer_guidelines) and update the checklist as you proceed. Please let me know if you have any questions.
