### 2017-02-02 #172 @braverock
@arfon I've completed the checklist.  I was previously familiar with the NSE package, so most of what I had to do was to check that JOSS guidelines had been met.  @keblu and @ArdiaD have implemented multiple ways of computing standard errors of a given metric, many of which have not appeared in the R language previously.  It is a worthwhile contribution to the literature, and I recommend publication without reservations.
