### 2017-02-04 #172 @ArdiaD
@keblu can you create a Zenodo account and create a DOI for the current nse project; please follow these steps: https://guides.github.com/activities/citable-code/
