### 2017-02-04 #172 @arfon
@braverock - many thanks for your review here ✨

@ArdiaD & @keblu your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00172 ⚡️ 🚀 💥
